refactor into modal component and reuse
